changed model to Config class and updated example
