+++ maps/config.py
@@ -1,117 +1,102 @@
+import geopandas as gpd
 import os
+from maps.topology import Topology
 
-# ROI
-# # padding arounf dtm to ensure reprojected dtm covers target area (in degrees)
-# step_out = 0.1
-# inset = 0  # unused??
 
-# # region of interest coordinates in metre-based system (or non-degree system)
-# minx = 500057
-# maxx = 603028
-# miny = 7455348
-# maxy = 7567953
-# model_top = 1200
-# model_base = -8200
+class Config(object):
+    def __init__(self, geology_file, fault_file, structure_file, mindep_file, bbox_3d, polygon, step_out, c_l={}):
+        self.bbox_3d = bbox_3d
+        self.bbox = tuple(list(bbox_3d.values())[:4])
+        self.polygon = polygon
+        self.step_out = step_out
+        self.c_l = c_l
 
-# PATHS
+        # Create file structure for model instance
+        self.geology_file = geology_file
+        self.structure_file = structure_file
+        self.fault_file = fault_file
+        self.mindep_file = mindep_file
 
-# flag to use local or WFS source for data inputs (True = local)
-# local_paths = True
+        self.project_path = './model-test'
 
-# test_data_path = './test_data/'
+        self.graph_path = self.project_path+'/graph'
+        self.tmp_path = self.project_path+'/tmp'
+        self.data_path = self.project_path+'/data'
+        self.dtm_path = self.project_path+'/dtm'
+        self.output_path = self.project_path+'/output'
+        self.vtk_path = self.project_path+'/vtk'
 
-# if(not os.path.isdir(test_data_path)):
-#     os.mkdir(test_data_path)
+        self.fault_file_csv = self.tmp_path + "/faults.csv"
+        self.structure_file_csv = self.tmp_path + "/structure.csv"
+        self.geology_file_csv = self.tmp_path + "/geology.csv"
+        self.mindep_file_csv = self.tmp_path + "/mindep.csv"
 
-geology_file = 'hams2_geol.shp'  # input geology file (if local)
-# input fault file (if local)
-fault_file = 'GEOS_GEOLOGY_LINEARSTRUCTURE_500K_GSD.shp'
-# input bedding orientation file (if local)
-structure_file = 'hams2_structure.shp'
-mindep_file = 'mindeps_2018.shp'  # input mineral deposit file (if local)
+        self.strat_graph_file = self.graph_path + "/graph_strat_NONE.gml"
+        self.dtm_file = self.dtm_path+'/dtm.tif'
+        self.dtm_reproj_file = self.dtm_path+'/dtm_rp.tif'
 
-# CRS
+        if(not os.path.isdir(self.project_path)):
+            os.mkdir(self.project_path)
+        if(not os.path.isdir(self.tmp_path)):
+            os.mkdir(self.tmp_path)
+        if(not os.path.isdir(self.data_path)):
+            os.mkdir(self.data_path)
+        if(not os.path.isdir(self.output_path)):
+            os.mkdir(self.output_path)
+        if(not os.path.isdir(self.dtm_path)):
+            os.mkdir(self.dtm_path)
+        if(not os.path.isdir(self.vtk_path)):
+            os.mkdir(self.vtk_path)
+        if(not os.path.isdir(self.graph_path)):
+            os.mkdir(self.graph_path)
 
-# coordinate reference system for imported dtms (geodetic lat/long WGS84)
-src_crs = {'init': 'EPSG:4326'}
-dst_crs = {'init': 'EPSG:28350'}  # coordinate system for data
+    def preprocess(self, command="plot"):
+        geology = gpd.read_file(self.geology_file, bbox=self.bbox)
+        lines = gpd.read_file(self.fault_file, bbox=self.bbox)
+        structures = gpd.read_file(self.structure_file, bbox=self.bbox)
+        mindep = gpd.read_file(self.mindep_file, bbox=self.bbox)
 
-# CODES AND LABELS
-# these refer to specific fields (codes) in GIS layer or database that contain the info needed for these calcs and text substrings (labels) in the contents of these fields
-c_l = {
-    # Orientations
-    "d": "DIP",  # field that contains dip information
-    "dd": "DIP_DIR",  # field that contains dip direction information
-    "sf": 'FEATURE',  # field that contains information on type of structure
-    # text to search for in field defined by sf code to show that this is a bedding measurement
-    "bedding": 'Bed',
-    # flag to determine measurement convention (currently 'strike' or 'dip direction')
-    "otype": 'dip direction',
-    # Stratigraphy
-    "g": 'GROUP_',  # field that contains coarser stratigraphic coding
-    "c": 'CODE',  # field that contains finer stratigraphic coding
-    "ds": 'DESCRIPTN',  # field that contains information about lithology
-    # field that contains alternate stratigraphic coding (not used??)
-    "u": 'UNITNAME',
-    "r1": 'ROCKTYPE1',  # field that contains  extra lithology information
-    "r2": 'ROCKTYPE2',  # field that contains even more lithology information
-    "sill": 'sill',  # text to search for in field defined by ds code to show that this is a sill
-    # text to search for in field defined by ds code to show that this is an intrusion
-    "intrusive": 'intrusive',
-    # text to search for in field defined by ds code to show that this is an volv=canic (not intrusion)
-    "volcanic": 'volcanic',
-    # Mineral Deposits
-    "msc": 'SITE_CODE',  # field that contains site code of deposit
-    "msn": 'SHORT_NAME',  # field that contains short name of deposit
-    "mst": 'SITE_TYPE_',  # field that contains site type of deposit
-    "mtc": 'TARGET_COM',  # field that contains target commodity of deposit
-    "mscm": 'SITE_COMMO',  # field that contains site commodity of deposit
-    "mcom": 'COMMODITY_',  # field that contains commodity group of deposit
-    # text to search for in field defined by mst code that shows site to ignore
-    "minf": 'Infrastructure',
-    # Timing
-    "min": 'MIN_AGE_MA',  # field that contains minimum age of unit defined by ccode
-    "max": 'MAX_AGE_MA',  # field that contains maximum age of unit defined by ccode
-    #faults and folds
-    "f": 'FEATURE',  # field that contains information on type of structure
-    # text to search for in field defined by f code to show that this is a fault
-    "fault": 'Fault',
-    # text to search for in field defined by f code to show that this is a fold axial trace
-    "fold": 'Fold axial trace',
-    # field that contains information on name of fault (not used??)
-    "n": 'NAME',
-    "t": 'TYPE',  # field that contains information on type of fold
-    # text to search for in field defined by t to show that this is a syncline
-    "syn": 'syncline',
-    # ids
-    "o": 'OBJECTID',  # field that contains unique id of geometry object
-    "gi": 'GEOPNT_ID'  # field that contains unique id of structure point
-}
+        if command == "plot":
+            try:
+                base = geology.plot(column=self.c_l['c'], figsize=(
+                    10, 10), edgecolor='#000000', linewidth=0.2)
+                structures.plot(ax=base, color='none', edgecolor='black')
+                lines.plot(ax=base, cmap='rainbow',
+                           column=self.c_l['f'], figsize=(10, 10), linewidth=0.4)
+                structures[['geometry', self.c_l['gi'],
+                            self.c_l['d'], self.c_l['dd']]].plot(ax=base)
+                self.polygon.plot(ax=base, color='none', edgecolor='black')
 
-# DECIMATION
+            except Exception as e:
+                print(e)
 
-# store every nth orientation (in object order) 0 = save all
-orientation_decimate = 0
-# store every nth contact point (in object order) 0 = save all
-contact_decimate = 10
-# store every nth fault point (in object order) 0 = save all
-fault_decimate = 5
-# store every nth fold axial trace point (in object order) 0 = save all
-fold_decimate = 5
+        elif command == "export_csv":
+            # Save geology polygons
+            hint_flag = False  # use GSWA strat database to provide relative age hints
+            sub_geol = geology[['geometry', self.c_l['o'], self.c_l['c'], self.c_l['g'],
+                                self.c_l['u'], self.c_l['min'], self.c_l['max'], self.c_l['ds'], self.c_l['r1'], self.c_l['r2']]]
+            Topology.save_geol_wkt(
+                sub_geol, self.geology_file_csv, self.c_l, hint_flag)
 
-# INTERPOLATION
+            # Save mineral deposits
+            sub_mindep = mindep[['geometry', self.c_l['msc'], self.c_l['msn'],
+                                 self.c_l['mst'], self.c_l['mtc'], self.c_l['mscm'], self.c_l['mcom']]]
+            Topology.save_mindep_wkt(
+                sub_mindep, self.mindep_file_csv, self.c_l)
 
-gridx = 50  # x grid dimensions (no of points, not distance) for interpolations
-gridy = 50  # x grid dimensions (no of points, not distance) for interpolations
-scheme = 'scipy_rbf'  # interpolation scheme
-# buffer distance for clipping points by faults (in metres or same units as dst_crs)
-dist_buffer = 5
-intrusion_mode = 0        # 1 all intrusions exluded from basal contacts, 0 only sills
-use_interpolations = False    # flag to use interpolated orientations or not.
+            # Save orientation data
+            sub_pts = structures[[
+                'geometry', self.c_l['gi'], self.c_l['d'], self.c_l['dd']]]
+            Topology.save_structure_wkt(
+                sub_pts, self.structure_file_csv, self.c_l)
 
-# ASSUMPTIONS
+            # Save faults
+            sub_lines = lines[['geometry', self.c_l['o'], self.c_l['f']]]
+            Topology.save_faults_wkt(sub_lines, self.fault_file_csv, self.c_l)
 
-pluton_dip = 45  # surface dip of pluton contacts
-# saucers: \__+_+__/  batholith: +/       \+   domes: /  +  + \  pendant: +\_____/+
-pluton_form = 'domes'
-fault_dip = 90  # surface dip of faults
+            # Return those sub frames for debugging
+            # return sub_geol, sub_lines, sub_pts, sub_mindep
+
+    def update_parfile(self):
+        Topology.save_parfile(self, self.c_l, self.data_path, self.geology_file_csv, self.fault_file_csv, self.structure_file_csv,
+                              self.mindep_file_csv, self.bbox[0], self.bbox[1], self.bbox[2], self.bbox[3], 500.0, 'Fe,Cu,Au,NONE')
