+++ setup.py
@@ -27,4 +27,3 @@ setuptools.setup(
     # ],
     python_requires='>=3.6',
 )
-
